bugfix: Fix lost messages with corrected generation of unique packet id.
The bug lies in the fact that the random number generator 'srand(time(NULL));' was called everytime a packet is builted. This cause the same numbers to be generated for some time, everytime the generator fct is called. The fix is to have the generator seeded once, at the beginning of the program. So in each test main file (one for server, and one for client), I seeded the generator just once. Now it generates a more unique id for every new message being sended.